clean up gallery editor context
